+++ test/dummy/app/models/user.rb
@@ -3,8 +3,4 @@ class User < ApplicationRecord
   # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :validatable
-
-  def full_name
-    [ first_name, last_name ].compact_blank.join(" ")
-  end
 end